Added preliminary organizations.html component, moved some helpers to utils/role_org.py, hooked it all up
