+++ routers/organization.py
@@ -5,10 +5,10 @@ from pydantic import BaseModel, ConfigDict, field_validator
 from sqlmodel import Session, select
 from utils.db import get_session
 from utils.auth import get_authenticated_user
-from utils.models import Organization, User, Role, UserOrganizationLink, ValidPermissions, RolePermissionLink, Permission, utc_time
+from utils.models import Organization, User, Role, UserOrganizationLink, ValidPermissions, utc_time
 from datetime import datetime
 from sqlalchemy import and_
-from typing import List
+from utils.role_org import get_organization, check_user_permission
 
 logger = getLogger("uvicorn.error")
 
@@ -102,128 +102,6 @@ class OrganizationUpdate(BaseModel):
         return cls(id=id, name=name)
 
 
-# -- Helper Functions --
-
-def get_user_organizations(
-    user_id: int,
-    session: Session,
-    include_deleted: bool = False
-) -> List[Organization]:
-    """
-    Retrieve all organizations a user is a member of.
-
-    Args:
-        user_id: ID of the user
-        session: Database session
-        include_deleted: Whether to include soft-deleted organizations
-
-    Returns:
-        List of Organization objects the user belongs to
-    """
-    query = (
-        select(Organization)
-        .join(UserOrganizationLink)
-        .where(UserOrganizationLink.user_id == user_id)
-    )
-
-    if not include_deleted:
-        query = query.where(Organization.deleted == False)
-
-    return list(session.exec(query))
-
-
-def get_organization(
-    org_id: int,
-    user_id: int,
-    session: Session,
-) -> Organization:
-    """
-    Retrieve a specific organization if the user is a member.
-
-    Args:
-        org_id: ID of the organization
-        user_id: ID of the user
-        session: Database session
-
-    Returns:
-        Organization object
-
-    Raises:
-        OrganizationNotFoundError: If organization doesn't exist
-        InsufficientPermissionsError: If user is not a member
-    """
-    # Check if user is a member of the organization
-    user_org = session.exec(
-        select(UserOrganizationLink).where(
-            and_(
-                UserOrganizationLink.user_id == user_id,
-                UserOrganizationLink.organization_id == org_id
-            )
-        )
-    ).first()
-
-    if not user_org:
-        raise InsufficientPermissionsError()
-
-    db_org = session.get(Organization, org_id)
-    if not db_org or db_org.deleted:
-        raise OrganizationNotFoundError()
-
-    return db_org
-
-
-def check_user_permission(
-    user_id: int,
-    org_id: int,
-    permission: ValidPermissions,
-    session: Session,
-) -> bool:
-    """
-    Check if user has the specified permission for the organization.
-
-    Args:
-        user_id: ID of the user
-        org_id: ID of the organization
-        permission: Permission to check
-        session: Database session
-
-    Returns:
-        True if user has permission, False otherwise
-    """
-    # Get user's role in the organization
-    user_org = session.exec(
-        select(UserOrganizationLink).where(
-            and_(
-                UserOrganizationLink.user_id == user_id,
-                UserOrganizationLink.organization_id == org_id
-            )
-        )
-    ).first()
-
-    if not user_org:
-        return False
-
-    # Get permission ID
-    permission_record = session.exec(
-        select(Permission).where(Permission.name == permission)
-    ).first()
-
-    if not permission_record:
-        return False
-
-    # Check if role has the permission
-    role_permission = session.exec(
-        select(RolePermissionLink).where(
-            and_(
-                RolePermissionLink.role_id == user_org.role_id,
-                RolePermissionLink.permission_id == permission_record.id
-            )
-        )
-    ).first()
-
-    return bool(role_permission)
-
-
 # -- Routes --
 
 @router.post("/", response_class=RedirectResponse)
@@ -268,7 +146,7 @@ def create_organization(
     session.add(user_org_link)
     session.commit()
 
-    return RedirectResponse(url=f"/organizations/{db_org.id}", status_code=303)
+    return RedirectResponse(url=f"/profile", status_code=303)
 
 
 @router.put("/{org_id}", response_class=RedirectResponse)
@@ -298,7 +176,7 @@ def update_organization(
     session.commit()
     session.refresh(organization)
 
-    return RedirectResponse(url=f"/organizations/{org.id}", status_code=303)
+    return RedirectResponse(url=f"/profile", status_code=303)
 
 
 @router.delete("/{org_id}", response_class=RedirectResponse)
@@ -318,4 +196,4 @@ def delete_organization(
     session.add(organization)
     session.commit()
 
-    return RedirectResponse(url="/organizations", status_code=303)
+    return RedirectResponse(url="/profile", status_code=303)
